+++ sdks/python/hatchet_sdk/opentelemetry/instrumentor.py
@@ -1,6 +1,8 @@
+import json
 from importlib.metadata import version
-from typing import Any, Callable, Collection, Coroutine
+from typing import Any, Callable, Collection, Coroutine, Union, cast
 
+from hatchet_sdk.contracts import workflows_pb2 as v0_workflow_protos
 from hatchet_sdk.utils.typing import JSONSerializableMapping
 
 try:
@@ -26,20 +28,29 @@ except (RuntimeError, ImportError, ModuleNotFoundError):
         "To use the HatchetInstrumentor, you must install Hatchet's `otel` extra using (e.g.) `pip install hatchet-sdk[otel]`"
     )
 
+import inspect
+from datetime import datetime
+
+from google.protobuf import timestamp_pb2
+
 import hatchet_sdk
 from hatchet_sdk import ClientConfig
 from hatchet_sdk.clients.admin import (
     AdminClient,
+    ScheduleTriggerWorkflowOptions,
     TriggerWorkflowOptions,
     WorkflowRunTriggerConfig,
 )
 from hatchet_sdk.clients.events import (
+    BulkPushEventOptions,
     BulkPushEventWithMetadata,
     EventClient,
     PushEventOptions,
 )
 from hatchet_sdk.contracts.events_pb2 import Event
+from hatchet_sdk.logger import logger
 from hatchet_sdk.runnables.action import Action
+from hatchet_sdk.utils.opentelemetry import OTelAttribute
 from hatchet_sdk.worker.runner.runner import Runner
 from hatchet_sdk.workflow_run import WorkflowRunRef
 
@@ -51,6 +62,13 @@ OTEL_TRACEPARENT_KEY = "traceparent"
 
 
 def create_traceparent() -> str | None:
+    logger.warning(
+        "As of SDK version 1.11.0, you no longer need to call `create_traceparent` manually. The traceparent will be automatically created by the instrumentor and injected into the metadata of actions and events when appropriate. This method will be removed in a future version.",
+    )
+    return _create_traceparent()
+
+
+def _create_traceparent() -> str | None:
     """
     Creates and returns a W3C traceparent header value using OpenTelemetry's context propagation.
 
@@ -71,6 +89,16 @@ def create_traceparent() -> str | None:
 def parse_carrier_from_metadata(
     metadata: JSONSerializableMapping | None,
 ) -> Context | None:
+    logger.warning(
+        "As of SDK version 1.11.0, you no longer need to call `parse_carrier_from_metadata` manually. This method will be removed in a future version.",
+    )
+
+    return _parse_carrier_from_metadata(metadata)
+
+
+def _parse_carrier_from_metadata(
+    metadata: JSONSerializableMapping | None,
+) -> Context | None:
     """
     Parses OpenTelemetry trace context from a metadata dictionary.
 
@@ -86,7 +114,7 @@ def parse_carrier_from_metadata(
     :Example:
 
     >>> metadata = {"traceparent": "00-0af7651916cd43dd8448eb211c80319c-b7ad6b7169203331-01"}
-    >>> context = parse_carrier_from_metadata(metadata)
+    >>> context = _parse_carrier_from_metadata(metadata)
     """
 
     if not metadata:
@@ -101,6 +129,16 @@ def parse_carrier_from_metadata(
 
 
 def inject_traceparent_into_metadata(
+    metadata: dict[str, str], traceparent: str | None = None
+) -> dict[str, str]:
+    logger.warning(
+        "As of SDK version 1.11.0, you no longer need to call `inject_traceparent_into_metadata` manually. The traceparent will automatically be injected by the instrumentor. This method will be removed in a future version.",
+    )
+
+    return _inject_traceparent_into_metadata(metadata, traceparent)
+
+
+def _inject_traceparent_into_metadata(
     metadata: dict[str, str], traceparent: str | None = None
 ) -> dict[str, str]:
     """
@@ -125,9 +163,8 @@ def inject_traceparent_into_metadata(
     >>> print(new_metadata)
     {"key": "value", "traceparent": "00-4bf92f3577b34da6a3ce929d0e0e4736-00f067aa0ba902b7-01"}
     """
-
     if not traceparent:
-        traceparent = create_traceparent()
+        traceparent = _create_traceparent()
 
     if not traceparent:
         return metadata
@@ -213,6 +250,14 @@ class HatchetInstrumentor(BaseInstrumentor):  # type: ignore[misc]
             self._wrap_async_run_workflow,
         )
 
+        ## IMPORTANT: We don't need to instrument the async version of `schedule_workflow`
+        ## because it just calls the sync version internally.
+        wrap_function_wrapper(
+            hatchet_sdk,
+            "clients.admin.AdminClient.schedule_workflow",
+            self._wrap_schedule_workflow,
+        )
+
         wrap_function_wrapper(
             hatchet_sdk,
             "clients.admin.AdminClient.run_workflows",
@@ -225,6 +270,19 @@ class HatchetInstrumentor(BaseInstrumentor):  # type: ignore[misc]
             self._wrap_async_run_workflows,
         )
 
+    def extract_bound_args(
+        self,
+        wrapped_func: Callable[..., Any],
+        args: tuple[Any, ...],
+        kwargs: dict[str, Any],
+    ) -> list[Any]:
+        sig = inspect.signature(wrapped_func)
+
+        bound_args = sig.bind(*args, **kwargs)
+        bound_args.apply_defaults()
+
+        return list(bound_args.arguments.values())
+
     ## IMPORTANT: Keep these types in sync with the wrapped method's signature
     async def _wrap_handle_start_step_run(
         self,
@@ -233,8 +291,11 @@ class HatchetInstrumentor(BaseInstrumentor):  # type: ignore[misc]
         args: tuple[Action],
         kwargs: Any,
     ) -> Exception | None:
-        action = args[0]
-        traceparent = parse_carrier_from_metadata(action.additional_metadata)
+        params = self.extract_bound_args(wrapped, args, kwargs)
+
+        action = cast(Action, params[0])
+
+        traceparent = _parse_carrier_from_metadata(action.additional_metadata)
 
         with self._tracer.start_as_current_span(
             "hatchet.start_step_run",
@@ -290,66 +351,293 @@ class HatchetInstrumentor(BaseInstrumentor):  # type: ignore[misc]
     ## IMPORTANT: Keep these types in sync with the wrapped method's signature
     def _wrap_push_event(
         self,
-        wrapped: Callable[[str, dict[str, Any], PushEventOptions | None], Event],
+        wrapped: Callable[[str, dict[str, Any], PushEventOptions], Event],
         instance: EventClient,
         args: tuple[
             str,
             dict[str, Any],
-            PushEventOptions | None,
+            PushEventOptions,
         ],
-        kwargs: dict[str, str | dict[str, Any] | PushEventOptions | None],
+        kwargs: dict[str, str | dict[str, Any] | PushEventOptions],
     ) -> Event:
+        params = self.extract_bound_args(wrapped, args, kwargs)
+
+        event_key = cast(str, params[0])
+        payload = cast(JSONSerializableMapping, params[1])
+        options = cast(
+            PushEventOptions,
+            params[2] if len(params) > 2 else PushEventOptions(),
+        )
+
+        attributes = {
+            OTelAttribute.EVENT_KEY: event_key,
+            OTelAttribute.ACTION_PAYLOAD: json.dumps(payload, default=str),
+            OTelAttribute.ADDITIONAL_METADATA: json.dumps(
+                options.additional_metadata, default=str
+            ),
+            OTelAttribute.NAMESPACE: options.namespace,
+            OTelAttribute.PRIORITY: options.priority,
+            OTelAttribute.FILTER_SCOPE: options.scope,
+        }
+
         with self._tracer.start_as_current_span(
             "hatchet.push_event",
+            attributes={
+                f"hatchet.{k.value}": v
+                for k, v in attributes.items()
+                if v
+                and k not in self.config.otel.excluded_attributes
+                and v != "{}"
+                and v != "[]"
+            },
         ):
-            return wrapped(*args, **kwargs)
+            options = PushEventOptions(
+                **options.model_dump(exclude={"additional_metadata"}),
+                additional_metadata=_inject_traceparent_into_metadata(
+                    dict(options.additional_metadata),
+                ),
+            )
+
+            return wrapped(event_key, dict(payload), options)
 
     ## IMPORTANT: Keep these types in sync with the wrapped method's signature
     def _wrap_bulk_push_event(
         self,
         wrapped: Callable[
-            [list[BulkPushEventWithMetadata], PushEventOptions | None], list[Event]
+            [list[BulkPushEventWithMetadata], BulkPushEventOptions], list[Event]
         ],
         instance: EventClient,
         args: tuple[
             list[BulkPushEventWithMetadata],
-            PushEventOptions | None,
+            BulkPushEventOptions,
         ],
-        kwargs: dict[str, list[BulkPushEventWithMetadata] | PushEventOptions | None],
+        kwargs: dict[str, list[BulkPushEventWithMetadata] | BulkPushEventOptions],
     ) -> list[Event]:
+        params = self.extract_bound_args(wrapped, args, kwargs)
+
+        bulk_events = cast(list[BulkPushEventWithMetadata], params[0])
+        options = cast(BulkPushEventOptions, params[1])
+
+        num_bulk_events = len(bulk_events)
+        unique_event_keys = {event.key for event in bulk_events}
+
         with self._tracer.start_as_current_span(
             "hatchet.bulk_push_event",
+            attributes={
+                "hatchet.num_events": num_bulk_events,
+                "hatchet.unique_event_keys": json.dumps(unique_event_keys, default=str),
+            },
         ):
-            return wrapped(*args, **kwargs)
+            bulk_events_with_meta = [
+                BulkPushEventWithMetadata(
+                    **event.model_dump(exclude={"additional_metadata"}),
+                    additional_metadata=_inject_traceparent_into_metadata(
+                        dict(event.additional_metadata),
+                    ),
+                )
+                for event in bulk_events
+            ]
+
+            return wrapped(
+                bulk_events_with_meta,
+                options,
+            )
 
     ## IMPORTANT: Keep these types in sync with the wrapped method's signature
     def _wrap_run_workflow(
         self,
-        wrapped: Callable[[str, Any, TriggerWorkflowOptions | None], WorkflowRunRef],
+        wrapped: Callable[
+            [str, JSONSerializableMapping, TriggerWorkflowOptions],
+            WorkflowRunRef,
+        ],
         instance: AdminClient,
-        args: tuple[str, Any, TriggerWorkflowOptions | None],
-        kwargs: dict[str, str | Any | TriggerWorkflowOptions | None],
+        args: tuple[str, JSONSerializableMapping, TriggerWorkflowOptions],
+        kwargs: dict[str, str | JSONSerializableMapping | TriggerWorkflowOptions],
     ) -> WorkflowRunRef:
+        params = self.extract_bound_args(wrapped, args, kwargs)
+
+        workflow_name = cast(str, params[0])
+        payload = cast(JSONSerializableMapping, params[1])
+        options = cast(
+            TriggerWorkflowOptions,
+            params[2] if len(params) > 2 else TriggerWorkflowOptions(),
+        )
+
+        attributes = {
+            OTelAttribute.WORKFLOW_NAME: workflow_name,
+            OTelAttribute.ACTION_PAYLOAD: json.dumps(payload, default=str),
+            OTelAttribute.PARENT_ID: options.parent_id,
+            OTelAttribute.PARENT_STEP_RUN_ID: options.parent_step_run_id,
+            OTelAttribute.CHILD_INDEX: options.child_index,
+            OTelAttribute.CHILD_KEY: options.child_key,
+            OTelAttribute.NAMESPACE: options.namespace,
+            OTelAttribute.ADDITIONAL_METADATA: json.dumps(
+                options.additional_metadata, default=str
+            ),
+            OTelAttribute.PRIORITY: options.priority,
+            OTelAttribute.DESIRED_WORKER_ID: options.desired_worker_id,
+            OTelAttribute.STICKY: options.sticky,
+            OTelAttribute.KEY: options.key,
+        }
+
         with self._tracer.start_as_current_span(
             "hatchet.run_workflow",
+            attributes={
+                f"hatchet.{k.value}": v
+                for k, v in attributes.items()
+                if v
+                and k not in self.config.otel.excluded_attributes
+                and v != "{}"
+                and v != "[]"
+            },
         ):
-            return wrapped(*args, **kwargs)
+            options = TriggerWorkflowOptions(
+                **options.model_dump(exclude={"additional_metadata"}),
+                additional_metadata=_inject_traceparent_into_metadata(
+                    dict(options.additional_metadata),
+                ),
+            )
+
+            return wrapped(workflow_name, payload, options)
 
     ## IMPORTANT: Keep these types in sync with the wrapped method's signature
     async def _wrap_async_run_workflow(
         self,
         wrapped: Callable[
-            [str, Any, TriggerWorkflowOptions | None],
+            [str, JSONSerializableMapping, TriggerWorkflowOptions],
             Coroutine[None, None, WorkflowRunRef],
         ],
         instance: AdminClient,
-        args: tuple[str, Any, TriggerWorkflowOptions | None],
-        kwargs: dict[str, str | Any | TriggerWorkflowOptions | None],
+        args: tuple[str, JSONSerializableMapping, TriggerWorkflowOptions],
+        kwargs: dict[str, str | JSONSerializableMapping | TriggerWorkflowOptions],
     ) -> WorkflowRunRef:
+        params = self.extract_bound_args(wrapped, args, kwargs)
+
+        workflow_name = cast(str, params[0])
+        payload = cast(JSONSerializableMapping, params[1])
+        options = cast(
+            TriggerWorkflowOptions,
+            params[2] if len(params) > 2 else TriggerWorkflowOptions(),
+        )
+
+        attributes = {
+            OTelAttribute.WORKFLOW_NAME: workflow_name,
+            OTelAttribute.ACTION_PAYLOAD: json.dumps(payload, default=str),
+            OTelAttribute.PARENT_ID: options.parent_id,
+            OTelAttribute.PARENT_STEP_RUN_ID: options.parent_step_run_id,
+            OTelAttribute.CHILD_INDEX: options.child_index,
+            OTelAttribute.CHILD_KEY: options.child_key,
+            OTelAttribute.NAMESPACE: options.namespace,
+            OTelAttribute.ADDITIONAL_METADATA: json.dumps(
+                options.additional_metadata, default=str
+            ),
+            OTelAttribute.PRIORITY: options.priority,
+            OTelAttribute.DESIRED_WORKER_ID: options.desired_worker_id,
+            OTelAttribute.STICKY: options.sticky,
+            OTelAttribute.KEY: options.key,
+        }
+
         with self._tracer.start_as_current_span(
             "hatchet.run_workflow",
+            attributes={
+                f"hatchet.{k.value}": v
+                for k, v in attributes.items()
+                if v
+                and k not in self.config.otel.excluded_attributes
+                and v != "{}"
+                and v != "[]"
+            },
         ):
-            return await wrapped(*args, **kwargs)
+            options = TriggerWorkflowOptions(
+                **options.model_dump(exclude={"additional_metadata"}),
+                additional_metadata=_inject_traceparent_into_metadata(
+                    dict(options.additional_metadata),
+                ),
+            )
+
+            return await wrapped(workflow_name, payload, options)
+
+    def _ts_to_iso(self, ts: Union[datetime, timestamp_pb2.Timestamp]) -> str:
+        if isinstance(ts, datetime):
+            return ts.isoformat()
+        elif isinstance(ts, timestamp_pb2.Timestamp):
+            return ts.ToJsonString()
+        else:
+            raise TypeError(f"Unsupported type for timestamp conversion: {type(ts)}")
+
+    ## IMPORTANT: Keep these types in sync with the wrapped method's signature
+    def _wrap_schedule_workflow(
+        self,
+        wrapped: Callable[
+            [
+                str,
+                list[Union[datetime, timestamp_pb2.Timestamp]],
+                JSONSerializableMapping,
+                ScheduleTriggerWorkflowOptions,
+            ],
+            v0_workflow_protos.WorkflowVersion,
+        ],
+        instance: AdminClient,
+        args: tuple[
+            str,
+            list[Union[datetime, timestamp_pb2.Timestamp]],
+            JSONSerializableMapping,
+            ScheduleTriggerWorkflowOptions,
+        ],
+        kwargs: dict[
+            str,
+            str
+            | list[Union[datetime, timestamp_pb2.Timestamp]]
+            | JSONSerializableMapping
+            | ScheduleTriggerWorkflowOptions,
+        ],
+    ) -> v0_workflow_protos.WorkflowVersion:
+        params = self.extract_bound_args(wrapped, args, kwargs)
+
+        workflow_name = cast(str, params[0])
+        schedules = cast(list[Union[datetime, timestamp_pb2.Timestamp]], params[1])
+        input = cast(JSONSerializableMapping, params[2])
+        options = cast(
+            ScheduleTriggerWorkflowOptions,
+            params[3] if len(params) > 3 else ScheduleTriggerWorkflowOptions(),
+        )
+
+        attributes = {
+            OTelAttribute.WORKFLOW_NAME: workflow_name,
+            OTelAttribute.RUN_AT_TIMESTAMPS: json.dumps(
+                [self._ts_to_iso(ts) for ts in schedules]
+            ),
+            OTelAttribute.ACTION_PAYLOAD: json.dumps(input, default=str),
+            OTelAttribute.PARENT_ID: options.parent_id,
+            OTelAttribute.PARENT_STEP_RUN_ID: options.parent_step_run_id,
+            OTelAttribute.CHILD_INDEX: options.child_index,
+            OTelAttribute.CHILD_KEY: options.child_key,
+            OTelAttribute.NAMESPACE: options.namespace,
+            OTelAttribute.ADDITIONAL_METADATA: json.dumps(
+                options.additional_metadata, default=str
+            ),
+            OTelAttribute.PRIORITY: options.priority,
+        }
+
+        with self._tracer.start_as_current_span(
+            "hatchet.schedule_workflow",
+            attributes={
+                f"hatchet.{k.value}": v
+                for k, v in attributes.items()
+                if v
+                and k not in self.config.otel.excluded_attributes
+                and v != "{}"
+                and v != "[]"
+            },
+        ):
+            options = ScheduleTriggerWorkflowOptions(
+                **options.model_dump(exclude={"additional_metadata"}),
+                additional_metadata=_inject_traceparent_into_metadata(
+                    dict(options.additional_metadata),
+                ),
+            )
+
+            return wrapped(workflow_name, schedules, input, options)
 
     ## IMPORTANT: Keep these types in sync with the wrapped method's signature
     def _wrap_run_workflows(
@@ -362,10 +650,37 @@ class HatchetInstrumentor(BaseInstrumentor):  # type: ignore[misc]
         args: tuple[list[WorkflowRunTriggerConfig],],
         kwargs: dict[str, list[WorkflowRunTriggerConfig]],
     ) -> list[WorkflowRunRef]:
+        params = self.extract_bound_args(wrapped, args, kwargs)
+        workflow_run_configs = cast(list[WorkflowRunTriggerConfig], params[0])
+
+        num_workflows = len(workflow_run_configs)
+        unique_workflow_names = {
+            config.workflow_name for config in workflow_run_configs
+        }
+
         with self._tracer.start_as_current_span(
             "hatchet.run_workflows",
+            attributes={
+                "hatchet.num_workflows": num_workflows,
+                "hatchet.unique_workflow_names": json.dumps(
+                    unique_workflow_names, default=str
+                ),
+            },
         ):
-            return wrapped(*args, **kwargs)
+            workflow_run_configs_with_meta = [
+                WorkflowRunTriggerConfig(
+                    **config.model_dump(exclude={"options"}),
+                    options=TriggerWorkflowOptions(
+                        **config.options.model_dump(exclude={"additional_metadata"}),
+                        additional_metadata=_inject_traceparent_into_metadata(
+                            dict(config.options.additional_metadata),
+                        ),
+                    ),
+                )
+                for config in workflow_run_configs
+            ]
+
+            return wrapped(workflow_run_configs_with_meta)
 
     ## IMPORTANT: Keep these types in sync with the wrapped method's signature
     async def _wrap_async_run_workflows(
@@ -378,10 +693,26 @@ class HatchetInstrumentor(BaseInstrumentor):  # type: ignore[misc]
         args: tuple[list[WorkflowRunTriggerConfig],],
         kwargs: dict[str, list[WorkflowRunTriggerConfig]],
     ) -> list[WorkflowRunRef]:
+        params = self.extract_bound_args(wrapped, args, kwargs)
+        workflow_run_configs = cast(list[WorkflowRunTriggerConfig], params[0])
+
         with self._tracer.start_as_current_span(
             "hatchet.run_workflows",
         ):
-            return await wrapped(*args, **kwargs)
+            workflow_run_configs_with_meta = [
+                WorkflowRunTriggerConfig(
+                    **config.model_dump(exclude={"options"}),
+                    options=TriggerWorkflowOptions(
+                        **config.options.model_dump(exclude={"additional_metadata"}),
+                        additional_metadata=_inject_traceparent_into_metadata(
+                            dict(config.options.additional_metadata),
+                        ),
+                    ),
+                )
+                for config in workflow_run_configs
+            ]
+
+            return await wrapped(workflow_run_configs_with_meta)
 
     def _uninstrument(self, **kwargs: InstrumentKwargs) -> None:
         self.tracer_provider = NoOpTracerProvider()